multilingual_ai_voice_assistant/requirements.txt
SpeechRecognition
pipwin
pyaudio
gTTS
google-genai
python-dotenv
streamlit